Documentation added

+++ model/activityLog.go
@@ -2,7 +2,7 @@ package model
 
 type ActivityLog struct {
 	UserId       string     `bson:"_id"`
-	Credits      int64      `bson:"credits";json:"credits"`
+	Credits      float64      `bson:"credits";json:"credits"`
 	SurveysMade  int64      `bson:"surveysMade";json:"surveysMade"`
 	SurveysTaken int64      `bson:"surveysTaken";json:"surveysTaken"`
 	Activities   []Activity `bson:"activities"`
